MARS+AQRUN

+++ commonCode/Well.vb
@@ -11,65 +11,65 @@ Friend Class Well
 	' Depth:        water depth (ft)
 	
 	Private mstrNameID As String
-	Private msngXg As Double
-	Private msngYg As Double
-	Private msngDepth As Double
-	
-	' properties
-	
-	
-	Public Property NameID() As String
-		Get
-			
-			NameID = mstrNameID
-			
-		End Get
-		Set(ByVal Value As String)
-			
-			mstrNameID = Value
-			
-		End Set
-	End Property
-	
-	
-	Public Property Xg() As Double
-		Get
-			
-			Xg = msngXg
-			
-		End Get
-		Set(ByVal Value As Double)
-			
-			msngXg = Value
-			
-		End Set
-	End Property
-	
-	
-	Public Property Yg() As Double
-		Get
-			
-			Yg = msngYg
-			
-		End Get
-		Set(ByVal Value As Double)
-			
-			msngYg = Value
-			
-		End Set
-	End Property
-	
-	
-	Public Property Depth() As Double
-		Get
-			
-			Depth = msngDepth
-			
-		End Get
-		Set(ByVal Value As Double)
-			
-			msngDepth = Value
-			
-		End Set
-	End Property
+    Private msngXg As Decimal
+    Private msngYg As Decimal
+    Private msngDepth As Decimal
+
+    ' properties
+
+
+    Public Property NameID() As String
+        Get
+
+            NameID = mstrNameID
+
+        End Get
+        Set(ByVal Value As String)
+
+            mstrNameID = Value
+
+        End Set
+    End Property
+
+
+    Public Property Xg() As Decimal
+        Get
+
+            Xg = msngXg
+
+        End Get
+        Set(ByVal Value As Decimal)
+
+            msngXg = Value
+
+        End Set
+    End Property
+
+
+    Public Property Yg() As Decimal
+        Get
+
+            Yg = msngYg
+
+        End Get
+        Set(ByVal Value As Decimal)
+
+            msngYg = Value
+
+        End Set
+    End Property
+
+
+    Public Property Depth() As Decimal
+        Get
+
+            Depth = msngDepth
+
+        End Get
+        Set(ByVal Value As Decimal)
+
+            msngDepth = Value
+
+        End Set
+    End Property
 End Class
